キャストが悪いのだろうか。調査

+++ FilterTimeFrameSliceByTrack.cxx
@@ -427,15 +427,15 @@ bool FilterTimeFrameSliceByTrack::RegisterDetectorConfig_Geometry()
       // Register KLDC
       if(DetectorName == "kldc"){
          std::cout << funcname << "Registering geometry for KLDC detector ID: " << detectorId << std::endl;
-         std::cout << "\t" << DetectorName << "(index: " << static_cast<uint8_t>(DetectorNameIdx) << "), " << PlaneName << "(index: " << static_cast<uint8_t>(PlaneNameIdx) << "), Segment: " << static_cast<uint8_t>(SegmentNumber) << std::endl;
+         std::cout << "\t" << DetectorName << "(index: " << static_cast<int>(DetectorNameIdx) << "), " << PlaneName << "(index: " << static_cast<int>(PlaneNameIdx) << "), Segment: " << static_cast<int>(SegmentNumber) << std::endl;
          for(int i=0+32; i<128-32; ++i){
             int ChannelNumber = i;
             uint32_t dopeKey_DETtoFE;
             #if 1
             std::cout << funcname << "channel map query: " << std::endl;
-            std::cout << "\t" << DetectorName << "(index: " << static_cast<uint8_t>(DetectorNameIdx) << "), " << PlaneName << "(index: " << static_cast<uint8_t>(PlaneNameIdx) << "), Segment: " << static_cast<uint8_t>(SegmentNumber) << ", ChannelName: " << ChannelName << "(index: " << static_cast<uint8_t>(ChannelNameIdx) << "), ChannelNumber: " << static_cast<uint16_t>(ChannelNumber) << std::endl;
+            std::cout << "\t" << DetectorName << "(index: " << static_cast<int>(DetectorNameIdx) << "), " << PlaneName << "(index: " << static_cast<int>(PlaneNameIdx) << "), Segment: " << static_cast<int>(SegmentNumber) << ", ChannelName: " << ChannelName << "(index: " << static_cast<int>(ChannelNameIdx) << "), ChannelNumber: " << static_cast<int>(ChannelNumber) << std::endl;
             #endif
-            bool found_DETtoFE = fChMap->getDopeKey_DETtoFE(DetectorNameIdx, PlaneNameIdx, static_cast<uint8_t>(SegmentNumber), ChannelNameIdx, static_cast<uint16_t>(ChannelNumber), dopeKey_DETtoFE);
+            bool found_DETtoFE = fChMap->getDopeKey_DETtoFE(DetectorNameIdx, PlaneNameIdx, SegmentNumber, ChannelNameIdx, static_cast<uint16_t>(ChannelNumber), dopeKey_DETtoFE);
             if(found_DETtoFE){
                chmap::FEAddrItem feaddritem = fChMap->getFEAddrItem(dopeKey_DETtoFE);
                uint32_t dopeKeyFEtoDET;
@@ -502,7 +502,7 @@ bool FilterTimeFrameSliceByTrack::RegisterDetectorConfig_DCTdcCalib()
       #endif
       if(isFirstEntry){
          std::cout << funcname << "Registering first tdccalib for KLDC detector ID: " << detectorId << std::endl;
-         std::cout << "\t" << DetectorName << ", " << PlaneName << ", Segment: " << static_cast<uint8_t>(SegmentNumber) << ",wire: " << wireId << std::endl;
+         std::cout << "\t" << DetectorName << ", " << PlaneName << ", Segment: " << static_cast<int>(SegmentNumber) << ",wire: " << wireId << std::endl;
          isFirstEntry = false;
       }
 
@@ -527,7 +527,7 @@ bool FilterTimeFrameSliceByTrack::RegisterDetectorConfig_DCTdcCalib()
       if(DetectorName == "kldc"){
          int ChannelNumber = wireId - 1; // wireId is 1-indexed, ChannelNumber is 0-indexed
          uint32_t dopeKey_DETtoFE;
-         bool found_DETtoFE = fChMap->getDopeKey_DETtoFE(DetectorName, PlaneName, static_cast<uint8_t>(SegmentNumber), ChannelName, static_cast<uint16_t>(ChannelNumber), dopeKey_DETtoFE);
+         bool found_DETtoFE = fChMap->getDopeKey_DETtoFE(DetectorName, PlaneName, SegmentNumber, ChannelName, static_cast<uint16_t>(ChannelNumber), dopeKey_DETtoFE);
          if(found_DETtoFE){
             chmap::FEAddrItem feaddritem = fChMap->getFEAddrItem(dopeKey_DETtoFE);
             uint32_t dopeKeyFEtoDET;
